archive/dsmAzureLib/requirements.txt
azure-storage-blob
azure-communication-email
setuptools >=65
twine >=4.0.2
wheel>=0.42.0